ui: ant drafts overflow fix
* closes #2431
