+++ ui/cap-react/src/antd/partials/FileList/containers/FileModal.js
@@ -1,0 +1,11 @@
+import { connect } from "react-redux";
+import FileModal from "../components/FileModal";
+
+const mapStateToProps = state => ({
+  versions: state.draftItem.get("fileVersions")
+});
+
+export default connect(
+  mapStateToProps,
+  null
+)(FileModal);
